feat(text): opt code
