+++ packages/core/src/2d/text/Font.ts
@@ -1,5 +1,7 @@
 import { RefObject } from "../../asset/RefObject";
 import { Engine } from "../../Engine";
+import { Texture2D } from "../../texture";
+import { CharInfoWithTexture } from "../assembler/CharInfo";
 import { FontAtlas } from "../atlas/FontAtlas";
 
 /**
@@ -26,7 +28,6 @@ export class Font extends RefObject {
   }
 
   private _name: string = "";
-  private _fontName: string = "";
   private _fontAtlasArray: Array<FontAtlas> = [];
 
   /**
@@ -37,14 +38,62 @@ export class Font extends RefObject {
   }
 
   /**
-   * The font name for Canvas.
+   * Add char into font atlas.
+   * @param id - The unique id for char
+   * @param imageSource - The source of texture
+   * @param width - The width of char
+   * @param height - The height of char
+   * @param offsetX - The char offset in X axis
+   * @param offsetY - The char offset in Y axis
+   * @param xAdvance - The next char start position in X axis
+   * @param ascent - The ascent of char
+   * @param descent - The descent of char
+   * @returns - The char's char info and texture
    */
-  get fontName(): string {
-    return this._fontName;
+  addCharInfo(
+    id: number,
+    imageSource: TexImageSource | OffscreenCanvas,
+    width: number,
+    height: number,
+    offsetX: number,
+    offsetY: number,
+    xAdvance: number,
+    ascent: number,
+    descent: number,
+  ): CharInfoWithTexture {
+    const { _fontAtlasArray: fontAtlasArray } = this;
+    if (fontAtlasArray.length === 0) {
+      this._createFontAtlas();
+    }
+
+    const lastIndex = fontAtlasArray.length - 1;
+    let lastFontAtlas = fontAtlasArray[lastIndex];
+    let charInfo = lastFontAtlas.addCharInfo(id, imageSource, width, height, offsetX, offsetY, xAdvance, ascent, descent);
+    if (!charInfo) {
+      lastFontAtlas = this._createFontAtlas();
+      charInfo = lastFontAtlas.addCharInfo(id, imageSource, width, height, offsetX, offsetY, xAdvance, ascent, descent);
+    }
+    return charInfo ? { charInfo: charInfo, texture: lastFontAtlas.texture } : null;
   }
 
-  set fontName(value: string) {
-    this._fontName = value;
+  /**
+   * Get char info.
+   * @param id - The unique id for char
+   * @returns - The char's char info and texture
+   */
+  getCharInfo(id: number): CharInfoWithTexture {
+    const { _fontAtlasArray: fontAtlasArray } = this;
+    for (let i = 0, l = fontAtlasArray.length; i < l; ++i) {
+      const fontAtlas = fontAtlasArray[i];
+      const charInfo = fontAtlas.getCharInfo(id);
+      if (charInfo) {
+        return {
+          charInfo: charInfo,
+          texture: fontAtlas.texture
+        }
+      }
+    }
+    return null;
   }
 
   /**
@@ -55,5 +104,14 @@ export class Font extends RefObject {
   private constructor(engine: Engine, name: string = "") {
     super(engine);
     this._name = name;
+  }
+
+  private _createFontAtlas(): FontAtlas {
+    const { engine } = this;
+    const fontAtlas = new FontAtlas(engine);
+    const texture = new Texture2D(engine, 512, 512);
+    fontAtlas.texture = texture;
+    this._fontAtlasArray.push(fontAtlas);
+    return fontAtlas;
   }
 }
